rag_service/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.6
pydantic-settings>=2.2
asyncpg
redis
rq
openai>=1.42
httpx
pandas
python-docx
pdfplumber
pdfminer.six
chardet
prometheus-client
python-multipart
tenacity
aiofiles
watchdog
pytesseract
pdf2image
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-asyncpg
opentelemetry-instrumentation-logging
cachetools>=5.3
numpy>=1.24
